Feat: refactor to ahhere to SOLID principles
